Refactoring, add CMake support, formatting changes
