+++ lib/basic/CMakeLists.txt
@@ -1,0 +1,38 @@
+cmake_minimum_required(VERSION 3.20)
+
+# Configure project
+project(
+    basic
+    DESCRIPTION "Basic lexer, parser and interpreter"
+    LANGUAGES C CXX
+    VERSION 1.0
+)
+
+# Build the main project executable
+add_library(${PROJECT_NAME}
+    "src/ast.c"
+    "src/basic_lexer.c"
+    "src/basic_parser.c"
+    "src/basic_program.c"
+    "src/basic_runner.c"
+    "src/utils.c"
+    "src/utils.h"
+)
+
+add_library(${PROJECT_NAME}::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
+
+# Add some preprocessor defines
+target_compile_definitions(${PROJECT_NAME} PUBLIC "$<$<CONFIG:DEBUG>:DEBUG>")
+
+add_dependencies(${PROJECT_NAME} basic_system data_structures)
+
+target_link_libraries(${PROJECT_NAME}
+    PRIVATE
+        basic_system::basic_system
+        data_structures::data_structures
+)
+
+target_include_directories(${PROJECT_NAME} PUBLIC
+	$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    PRIVATE ${CMAKE_CURRENT_BINARY_DIR}
+)
