+++ lib/basic/include/basic/basic.h
@@ -1,5 +1,4 @@
-#ifndef _BASIC_H
-#define _BASIC_H
+#pragma once
 
 /* BASIC Program parser / interpreter common header file */
 
@@ -14,19 +13,19 @@
  * */
 
 #include "ast.h"
-#include "stack.h"
+#include <data_structures/stack.h>
 
 /* For printing Log messages */
 extern unsigned int log_print_mask;
 
-//Log level mask
+// Log level mask
 #define LOGTYPE_DEBUG   (1 << 0)
 #define LOGTYPE_INFO    (1 << 1)
 #define LOGTYPE_MESSAGE (1 << 2)
 #define LOGTYPE_ERROR   (1 << 3)
 #define LOGMASK_ALL     0xffff
 
-//Printf modified to allow only selected messages
+// Printf modified to allow only selected messages
 void lprintf(const char *tag, unsigned int level_mask, const char *format, ...);
 
 /* AST parser (tokenizer) */
@@ -49,27 +48,27 @@ typedef struct {
     char *token_at;
 } BASICToken;
 
-//Structure with all the stuff needed to convert a BASIC program to an AST
+// Structure with all the stuff needed to convert a BASIC program to an AST
 typedef struct {
-    //List of tokens
+    // List of tokens
     BASICToken *tokens;
     int tokens_length;
-    //Stack for knowning current scope
+    // Stack for knowning current scope
 } BASICParseTree;
 
 typedef struct {
     char *program_source;
     BASICParseTree program_tokens;
     ASTNode *program_sequence;
-    //Map between line number and which instruction to execute on that line. For non-linear control flow
-    //BASICLineNode program_line_instruction;
+    // Map between line number and which instruction to execute on that line. For non-linear control flow
+    // BASICLineNode program_line_instruction;
 } BASICProgram;
 
 BASICProgram *basic_create_program();
 void basic_clear_program(BASICProgram *program);
 void basic_destroy_program(BASICProgram *program);
 
-//Recognized tokens
+// Recognized tokens
 extern char *PARSE_KEYWORDS[];
 extern int PARSE_KW_COUNT;
 extern char *PARSE_BOOLEAN[];
@@ -77,7 +76,7 @@ extern char PARSE_OPERATORS[];
 extern char PARSE_SEPARATOR[];
 extern char PARSE_WS_CHAR[];
 
-//Index of the keyword in the above PARSE_KEYWORDS array
+// Index of the keyword in the above PARSE_KEYWORDS array
 #define KEYWORD_IDX_WHILE       0
 #define KEYWORD_IDX_IF          1
 #define KEYWORD_IDX_THEN        2
@@ -85,10 +84,10 @@ extern char PARSE_WS_CHAR[];
 #define KEYWORD_IDX_END         4
 #define KEYWORD_IDX_GOTO        5
 
-//Lexer
+// Lexer
 int basic_tokenize(BASICProgram *program);
 
-//Parser
+// Parser
 int basic_parse_form_expression(BASICParseTree *ptree, ASTNode *root, int parse_from, int parse_to, int *parse_new_pos);
 int basic_parse_form_function(BASICParseTree *ptree, ASTNode *root, int parse_from, int parse_to, int *parse_new_pos);
 
@@ -138,11 +137,11 @@ void basic_var_assignment(BASICRuntime *runtime, ASTNode *args);
 void basic_init_constants(BASICRuntime *runtime);
 KeywordAction basic_evaluate_keyword_block(BASICRuntime *runtime, ASTNode *pc, ASTNode **nextpc);
 
-//Keyword handlers
+// Keyword handlers
 KeywordAction basic_eval_kw_if(BASICRuntime *runtime, ASTNode *pc, ASTNode **nextpc);
 KeywordAction basic_eval_kw_while(BASICRuntime *runtime, ASTNode *pc, ASTNode **nextpc);
 
-//Built-in functions
+// Built-in functions
 ASTNodeData _basic_fn_print(BASICRuntime *runtime, ASTNode *args);
 ASTNodeData _basic_fn_max(BASICRuntime *runtime, ASTNode *args);
 ASTNodeData _basic_fn_min(BASICRuntime *runtime, ASTNode *args);
@@ -151,5 +150,3 @@ ASTNodeData _basic_fn_toint(BASICRuntime *runtime, ASTNode *arg);
 ASTNodeData _basic_fn_toflt(BASICRuntime *runtime, ASTNode *arg);
 ASTNodeData _basic_fn_rand(BASICRuntime *runtime, ASTNode *arg);
 ASTNodeData _basic_fn_irand(BASICRuntime *runtime, ASTNode *arg);
-
-#endif
