+++ lib/basic/src/basic_program.c
@@ -1,19 +1,19 @@
 
-#include "basic.h"
+#include "basic/basic.h"
 
-//Standard libraries
+// Standard libraries
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
 
 unsigned int log_print_mask = -1;
 void lprintf(const char *tag, unsigned int level_mask, const char *format, ...) {
-    //Infinite arguments
+    // Infinite arguments
     va_list args;
     va_start(args, format);
     if ((log_print_mask & level_mask) != 0) {
         printf("(%s) ", tag);
-        //printf, but with variable arguments list
+        // printf, but with variable arguments list
         vprintf(format, args);
     }
     va_end(args);
@@ -42,18 +42,18 @@ BASICProgram *basic_create_program() {
 }
 
 void basic_clear_program(BASICProgram *program) {
-    //Delete each child node and it's children recursively
+    // Delete each child node and it's children recursively
     ast_delete_children_cascade(program->program_sequence);
 }
 
 void basic_destroy_program(BASICProgram *program) {
-    //Delete all the nodes in the program sequence
+    // Delete all the nodes in the program sequence
     basic_clear_program(program);
-    //Delete the sequence node itself
+    // Delete the sequence node itself
     ast_delete_node(program->program_sequence);
-    //Delete Parse tree data
+    // Delete Parse tree data
     if (program->program_tokens.tokens != NULL) free(program->program_tokens.tokens);
-    //Finally delete the program object
+    // Finally delete the program object
     free(program);
 }
