+++ lib/basic_system/include/basic_system/system.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void system_sleep(unsigned long long us);
+int system_random_int();
+float system_random_float();
