+++ lib/basic_system/src/hal.h
@@ -1,0 +1,3 @@
+#pragma once
+
+
