+++ lib/data_structures/include/data_structures/stack.h
@@ -1,5 +1,4 @@
-#ifndef _STACK_H
-#define _STACK_H
+#pragma once
 
 typedef struct _stk_node {
     void *data;
@@ -11,5 +10,3 @@ void stack_delete_node(StackNode *node);
 void stack_clear(StackNode **top);
 void stack_push(StackNode **top, StackNode *node);
 StackNode *stack_pop(StackNode **top);
-
-#endif
