+++ src/utils.h
@@ -1,0 +1,5 @@
+#pragma once
+
+// Utility functions
+
+void print_level_space(int level);
